requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
Markdown
pillow
pipenv
platformdirs
setuptools
sqlparse
tzdata
virtualenv